--- lib/core/animation/animation.dart
@@ -1,1 +1,0 @@
-
