+++ lib/models/speaker_model.dart
