+++ lib/widgets/about_section.dart
