+++ lib/widgets/animated_banner_widget.dart
