+++ lib/widgets/leyends_widget.dart
